requirements.txt
anyio==3.6.2
asgiref==3.6.0
click==8.1.3
fastapi==0.68.2
h11==0.14.0
httptools==0.5.0
idna==3.4
mysqlclient==2.1.1
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.14.2
typing_extensions==4.5.0
uvicorn==0.15.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4

mysql-connector-python~=8.0.32